requirements.txt
streamlit>=1.39.0
streamlit-lottie
pandas>=2.2.2
pdfplumber>=0.11.4
docx2txt>=0.8
scikit-learn>=1.5.2
sentence-transformers>=3.0.1
nltk>=3.9.1
torch>=2.2.0
transformers>=4.42.0

numpy>=1.26.4

# Web Scraping & Automation
beautifulsoup4>=4.12.0
selenium>=4.15.0
requests>=2.31.0
webdriver-manager>=4.0.0

# AI/ML & NLP
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
spacy>=3.7.0
textblob>=0.17.1

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0

# Utilities
python-dotenv>=1.0.0
APScheduler>=3.10.0
PyGithub>=2.1.1
google-api-python-client>=2.100.0
